fix(adapters): correct Destatis GENESIS against the live API
Ran every endpoint against a real GENESIS account, which settled the three
that shipped unverified — data/table, metadata/table and catalogue/statistics
all answer Code 0 — and contradicted three things the adapter claimed.

The documented "silent guest fallback" does not exist. find/find is a public
service: it returns the same 80 tables with or without credentials, so an
empty search means nothing was found, never that authentication failed.
Telling an agent otherwise sends it chasing a credential problem that isn't
there. The real split is that data, metadata and catalogue are gated (Code 15
when credentials never arrive, Code 2 when they arrive wrong).

catalogue/statistics filters by statistic CODE, not by keyword: the parameter
described as a "filter keyword" returns Code 104 and an empty list for a word
like "Bevoelkerung", while "122*" returns four. It is now named `selection`
and described as the code pattern it is. Pinning searchcriterion/sortcriterion
also turns the permanent Code 22 warning into a clean Code 0.

logincheck echoes the caller's API token back verbatim in `Username`, where
every other endpoint masks it. That would reach the model's context and be
persisted in tool_invocations.output, so a responseMapping reduces it to
`isGuest`, which is the only part anyone needed.

The connector now pins the UTF-8 charset on its form bodies, without which
umlaut searches silently return nothing.

The static guards move into the live spec alongside the other 147 adapter
specs, and the live half now drives the shipped endpointMapping blocks rather
than hand-written copies, so a mapping that drifts from what was verified
fails here instead of in a customer's connector. Verified: 16/16 green
against the real API.

Guides in en/de/it carry the corrected explanation, plus a note that an older
install must be removed and re-installed — catalogue re-sync updates tools but
never a connector's auth type or base URL.
