+++ content/guides/de/destatis-genesis-to-mcp.mdx
@@ -30,7 +30,9 @@ Destatis empfiehlt die Identifizierung über einen **persönlichen API-Token** a
 
 Öffnen Sie `http://localhost:3000/connectors/store` und klicken Sie auf **Import**. Tragen Sie **entweder** Ihren API-Token **oder** Ihre Nutzerkennung in das Feld *Username or Token* ein. Das Feld *Password* ist optional — bei Verwendung eines Tokens lassen Sie es leer.
 
-> **Hinweis zur Authentifizierung:** Destatis hat GET-Anfragen mit Zugangsdaten in der URL zum **30.06.2025** endgültig abgeschaltet. Die API ist seitdem ausschließlich per POST mit Zugangsdaten im HTTP-Header erreichbar; der Adapter ist entsprechend konfiguriert. Kommt eine Tabellensuche leer zurück, rufen Sie zuerst `destatis_login_check` auf: Lassen sich die Zugangsdaten nicht auflösen, antwortet `find/find` weiterhin mit `Code: 0` als Gast-Konto `GAST` — alle Trefferlisten sind dann `null`, was von einer echten Nulltreffer-Suche nicht zu unterscheiden ist. Die Daten- und Metadaten-Dienste scheitern dagegen deutlich mit `Code: 15`.
+> **Hinweis zur Authentifizierung:** Destatis hat GET-Anfragen mit Zugangsdaten in der URL zum **30.06.2025** endgültig abgeschaltet. Die API ist seitdem ausschließlich per POST mit Zugangsdaten im HTTP-Header erreichbar; der Adapter ist entsprechend konfiguriert. `find/find` ist ein **öffentlicher** Dienst: Er antwortet mit `Code: 0` und denselben Treffern, ob Zugangsdaten mitgesendet werden oder nicht. Eine leere Tabellensuche bedeutet also, dass wirklich nichts gefunden wurde, und ist nie ein Hinweis auf ein Zugangsdaten-Problem. Die Daten-, Metadaten- und Katalogdienste sind dagegen geschützt: `Code: 15` heißt, die Zugangsdaten sind nicht angekommen, `Code: 2`, dass sie angekommen, aber falsch sind. Welches Konto antwortet, zeigt `destatis_login_check` — bei der anonymen Kennung `GAST` meldet es `isGuest: true`.
+
+> **Bestehende Installation aktualisieren:** Entfernen Sie den Connector und installieren Sie ihn neu aus dem Store, statt ihn zu synchronisieren. Die Katalog-Synchronisierung aktualisiert die Tool-Definitionen, nicht aber Authentifizierungsart und Basis-URL eines Connectors — ein synchronisierter alter Connector würde seine Zugangsdaten weiterhin als URL-Parameter an den abgeschalteten Host senden.
 
 ### Schritt 4: Mit KI-Agent verbinden
